Move memcached store to a separate package
